Update ci.yml
In this modification, I've replaced the export command with echo "AWS_DEFAULT_REGION=us-east-1" >> $GITHUB_ENV to set the environment variable for the entire workflow. Additionally, I've added the region: us-east-1 parameter to the aws-actions/amazon-ecr-login@v2 action to explicitly specify the region. This should address the "Region is missing" error in your workflow.